Boot32
I've started to work more on Boot32, this is so I can setup
an IDT and a PML4 page map so that I can enter 64 bit mode.
Although I am questioning the structure of how I want to set it up, and
what else I would like to do in 32 bit mode.