Changed some things
I've updated things, I really don't like having to jump
to a memory address and assuming that there would be
code there, but I'm not sure of other ways to do it.
If I find another way to do it which does not involve jumping to a
memory address, I will implement it.
For now I've changed some names, and makefiles.