Paging update
I was getting the format of virtual addresses wrong, I've fixed that.
Now I know that the framebuffer is a different page directory
and a different page table.
I still cannot write to the framebuffer after paging is enabled.
As well as that in memdump.bin the memory address of the framebuffer is
stored at 0x14000, and I'm not sure why.
Paging is very difficult.