I nformation transfer
The transfer of information from the bootloader to
the kernel is still very broken.
I am unsure of how to fix it, but I'll keep trying.
My current guesses are that the stack space is not proper or the
address transfer through registers do not set the percise location,
the register isn't zeroed before putting the address into the register.